Eliminated the required use of AWS key and secret as authentication
The goal here is to use IAM role permissions on an AWS instance to allow
writes to S3.  Explicitly passing in a key and secret short-cicuits
the boto3 permission chain as described by AWS:
(https://boto3.amazonaws.com/v1/documentation/api/latest/guide/configuration.html)

'''
The mechanism in which boto3 looks for credentials is to search through
a list of possible locations and stop as soon as it finds credentials.
The order in which Boto3 searches for credentials is:
    Passing credentials as parameters in the boto.client() method
    Passing credentials as parameters when creating a Session object
    Environment variables
    Shared credential file (~/.aws/credentials)
    AWS config file (~/.aws/config)
    Assume Role provider
    Boto2 config file (/etc/boto.cfg and ~/.boto)
    Instance metadata service on an Amazon EC2 instance that has an IAM
role configured.
'''

Want to use the last method, instance metadata, which is made possible
by removing explicit use of secret and key.